Create README.md
# smart-mail-transfer-smart-contract
The Smart Mail Transfer System is a decentralized application (DApp) built on the Ethereum blockchain using a smart contract written in Solidity. This system facilitates the secure and transparent transfer of virtual mail from a sender to a receiver. Each mail item is assigned a unique identification number, and the transfer process is executed through a trustless smart contract.